move optimizer config to base class
